+++ blindr/api/pictures.py
@@ -1,5 +1,5 @@
 from flask.ext import restful
-from authenticate import authenticate
+from .authenticate import authenticate
 from flask import request, abort, jsonify
 import time
 from itertools import permutations
@@ -7,9 +7,9 @@ from sqlalchemy.exc import IntegrityError
 import facebook
 
 import config
-from models.event import Event
-from models.user import User
-from models.match import Match
+from blindr.models.event import Event
+from blindr.models.user import User
+from blindr.models.match import Match
 
 class Pictures(restful.Resource):
     method_decorators=[authenticate]
