+++ blindr/data/adjectives.txt
