+++ blindr/utils.py
