+++ boto.cfg
@@ -1,0 +1,4 @@
+[Credentials]
+aws_access_key_id = <your_access_key_here>
+aws_secret_access_key = <your_secret_key_here>
+
