--- database/setup.py
@@ -1,4 +1,0 @@
-from models import *
-import config
-
-Base.metadata.create_all(config.engine)
