+++ run.py
@@ -1,0 +1,2 @@
+from blindr import app
+app.run()
