+++ setup.py
@@ -1,0 +1,2 @@
+from blindr import db
+db.create_all()
