+++ shell.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+import os
+import readline
+from pprint import pprint
+
+from flask import *
+from blindr import *
+
+os.environ['PYTHONINSPECT'] = 'True'
